Introduce a new TTL for the ephemeral (queue) FIFO locks
Leverage Extend capabilities from Extendable Redlock.
Additionally, the autoextending thread feature from Extendable Redlock is gone.
If someone requires this, it needs to be in their own application. Currently, it is not
reasonable to force the user to call is_valid on a lock. If you were issued the lock,
you must intentionally refresh it to keep it, and check the returned data.